refactor(schema-analyser): schema analyser is now its own class
